change default for JsonSymbol deserializer for lists to attempt the [type,value]
and make DSL Predicate deserialization the default for predicates
